refactor(cli): split shared CLI helpers into a cli module
`lsp_impl::cli_format` had grown to hold the CLI session lifecycle shared by
BOTH subcommands — `cli_initialize`, `cli_shutdown`, `cli_can_handle_path`, and
the cold-start readiness waits — alongside the format-specific wrapper. The name
no longer matched the contents, and `diagnose` silently depended on a file
called "format" for its shared plumbing.

Reorganize into a `cli` module mirroring `crate::cli`'s format/diagnose split:
`cli.rs` is a thin module root; `cli/session.rs` holds the shared `impl Kakehashi`
session lifecycle; `cli/format.rs` keeps only `cli_format_text`; `cli/diagnose.rs`
keeps only `cli_diagnose_text`.

Behavior-preserving move — no logic change; the full lib suite (2125 tests) is
unchanged.
