+++ src/lsp/lsp_impl.rs
@@ -1,6 +1,5 @@
 pub(crate) mod bridge_context;
-mod cli_diagnose;
-mod cli_format;
+mod cli;
 mod coordinator;
 pub(crate) use coordinator::DiagnosticPublisher;
 mod kakehashi;
